refactor(api): tighten ConfigKV cache boundary
