+++ server/apps/api/src/services/adapters/config-kv/contracts.test.ts
@@ -1,0 +1,21 @@
+import { describe, expect, it } from 'vitest'
+
+import { parseConfigKVInvalidation } from './contracts'
+
+describe('configKV invalidation contract', () => {
+  it('accepts a declared ConfigKV key', () => {
+    expect(parseConfigKVInvalidation(JSON.stringify({
+      key: 'FLUX_PER_REQUEST',
+      version: 1,
+      publishedAt: 1,
+    }))).toMatchObject({ key: 'FLUX_PER_REQUEST' })
+  })
+
+  it('rejects an unknown ConfigKV key', () => {
+    expect(() => parseConfigKVInvalidation(JSON.stringify({
+      key: 'UNKNOWN_CONFIG_KEY',
+      version: 1,
+      publishedAt: 1,
+    }))).toThrow('ConfigKV invalidation key is unknown')
+  })
+})
